STYLE: Refactor code removing vtkMRMLChestRTColorTableNode class
This class was a quasi complete copy of vtkMRMLColorTableNode where
a custom type with 256 color was added.

This commit simplifies the overall maintenance by removing the classes.
Instead a color table of type user is added when the  RegionTypeDisplay
logic is associated with a MRML scene.

Note that similarly to the previous implementation no color are added
and it initially contains 256 colors set to 0.