Refactor CIP_ParenchymaAnalysis module
Module logic now does not need labelmap for any additional report generation after the module widget's Apply method is completed.
This way when the input is a segmentation node, the temporary labelmap node does not have to be kept after the Apply method.

Also renamed variables to be more consistent across the file and simplified region color management (retrieve from the labelmap input and save it in a dictionary).