COMP: Fix VTK-9.1 compatibility
QVTKOpenGLNativeWidget::GetRenderWindow was renamed to QVTKOpenGLNativeWidget::renderWindow() some time ago,
but the old method name was still kept, so it was not an issue that SlicerCIP still used the old name.

However, the old method name was removed in VTK-9.1, therefore we need to switch to that in recent VTK versions.